requirements.txt
asgiref==3.5.2
django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
pillow==9.2.0
markdown==3.4.1
django-filter==22.1
